+++ Enum.lua
@@ -100,6 +100,6 @@ Private.Enum.FeatureFlag = {
 	IgnoreFriendlies = 11,
 	UnfillChannels = 12,
 	HideWhenUninterruptible = 13,
-	PlaySoundOnCastStart = 14,
+	PlayTTSOnCastStart = 14,
 	ShowAvailableInterrupts = 15,
 }